organizing and splitting functions
