+++ index.js
@@ -1,331 +1,19 @@
-// =============================== Strings =========================== //
-
-/**
- * Returns a capitalized String.
- * @param {string} s String that will be Capitalized.
- */
-const capitalize = (s) => {
-  if (typeof s !== "string") return "";
-  return s.charAt(0).toUpperCase() + s.slice(1);
-};
-
-/**
- * Returns a truncated string.
- * @param {string} text String to be truncated.
- * @param {number} num Max length of the `String` that will be truncated.
- */
-const truncate = (text, num = 10) => {
-  if (text.length > num) {
-    return `${text.substring(0, num - 3)}...`;
-  }
-  return text;
-};
-
-/**
- * Returns toggled '1' or '0'.
- * @param {string} strNum String number "0" or "1".
- */
-const toggleStrNum = (strNum) => {
-  if (strNum === "0" || strNum === "1") return strNum === "0" ? "1" : "0";
-  return null;
-};
-
-/**
- * Converts CamelCase string into string with spaces
- * @param {string} str CamelCase string
- * @returns {string}
- */
-const replaceCamelCaseWithSpaces = (str) =>
-  str.replace(/\B[A-Z]\B/g, (match) => ` ${match}`);
-
-/**
- * Logging formatted strings.
- * @param {any} input
- */
-const logFormattedStrings = (input) =>
-  console.log(JSON.stringify(input, null, 4));
-
-// =============================== Numbers =========================== //
-
-/**
- * Returns a random number between min to max.
- * @param {number} min Min number [default = 0].
- * @param {number} max Max Number [default = 1000].
- *
- */
-const randomNumber = (min = 0, max = 1000) =>
-  Math.ceil(min + Math.random() * (max - min));
-
-/**
- * Returns Converting Bytes in a Readable Human File Sizes.
- * @param {number} bytes Bytes in number.
- */
-const humanFileSize = (bytes) => {
-  let BYTES = bytes;
-  const thresh = 1024;
-
-  if (Math.abs(BYTES) < thresh) {
-    return `${BYTES} B`;
-  }
-
-  const units = ["KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"];
-
-  let u = -1;
-  const r = 10 ** 1;
-
-  do {
-    BYTES /= thresh;
-    u += 1;
-  } while (
-    Math.round(Math.abs(BYTES) * r) / r >= thresh &&
-    u < units.length - 1
-  );
-
-  return `${BYTES.toFixed(1)} ${units[u]}`;
-};
-
-/**
- * Returns a formatted number with separators based on Options.
- * @param {number} num integer.
- * @param {string} lang visit => https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString#using_locales
- * @param {object} options visit => https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString#using_options
- */
-const formatNumber = (num, lang, options) =>
-  num.toLocaleString(lang, { ...options });
-
-// =============================== Arrays =========================== //
-
-/**
- * Check that every element in an array exists in the other array.
- * @param {array} baseArr The array to make sure it has all the values.
- * @param {arr} arr The other array that will be compared with.
- */
-const containsAll = (baseArr, arr) => arr.every((ele) => baseArr.includes(ele));
-
-/**
- * Returns a unique array of objects based on a key.
- * @param {array} array Array of objects.
- * @param {string} key A unique property of each object [default = 'id'].
- */
-const getUniqueObjs = (array, key = "id") => {
-  const ids = [];
-  const output = [];
-  array.forEach((ele) => {
-    if (!ids.includes(ele[key])) {
-      ids.push(ele[key]);
-      output.push(ele);
-    }
-  });
-  return output;
-};
-// =============================== Time =========================== //
-
-/**
- * Getting an Array of Times + "AM" or "PM".
- * @param {number} startTime Starting hour [in minutes].
- * @param {number} minutesInterval every *n* minutes [default = 15].
- */
-const getTimes = (startTime = 0, minutesInterval = 15) => {
-  const times = []; // time array
-  const x = minutesInterval; // minutes interval
-  let tt = startTime; // start time
-  const ap = ["AM", "PM"]; // AM-PM
-
-  const formatHrs = (hh) => {
-    const _12Format = hh === 12 ? 12 : hh % 12;
-    return _12Format === 0 ? 12 : _12Format;
-  };
-
-  // loop to increment the time and push results in array
-  for (let i = 0; tt < 24 * 60; i += 1) {
-    const hh = Math.floor(tt / 60); // getting hours of day in 0-24 format
-    const mm = tt % 60; // getting minutes of the hour in 0-55 format
-
-    const hours = `0${formatHrs(hh)}`.slice(-2);
-    const minutes = `0${mm}`.slice(-2);
-    const AMPM = ap[Math.floor(hh / 12)];
-
-    times[i] = `${hours}:${minutes} ${AMPM}`; // pushing data in array in [00:00 - 12:00 AM/PM format]
-    tt += x;
-  }
-  return times;
-};
-
-// =============================== Objects =========================== //
-
-/**
- * Convert Objects to Form Data Format.
- * @param {object} obj
- */
-const toFormData = (obj) => {
-  const formBody = new FormData();
-  Object.keys(obj).forEach((key) => {
-    if (Array.isArray(obj[key])) {
-      obj[key].forEach((val, i) => {
-        formBody.append(`${key}[${i}]`, val);
-      });
-    } else formBody.append(key, obj[key]);
-  });
-  return formBody;
-};
-
-/**
- * Soft  cloning objects
- * @param {object} obj Input
- */
-const softClone = (obj) => JSON.parse(JSON.stringify(obj));
-
-// =============================== HTML =========================== //
-
-/**
- * Getting the inner `Text` of an `HTML` string.
- * @param {string} str A string of HTML.
- */
-const getInnerHTML = (str) => str.replace(/(<([^>]+)>)/gi, "").trim();
-
-// =============================== Window =========================== //
-
-/**
- * Detect screen width and returns a string representing the width of the screen.
- */
-const getScreenWidth = () => {
-  const screenWidth = window.screen.width;
-  if (screenWidth <= 425) return "mobile";
-  if (screenWidth <= 768) return "tablet";
-  if (screenWidth <= 1024) return "laptopSm";
-  if (screenWidth <= 1440) return "laptopLg";
-  if (screenWidth <= 2560) return "HD";
-  return screenWidth;
-};
-
-/**
- * Scroll to top
- */
-const toTop = (behavior = "smooth") => {
-  window.scroll({ top: 0, left: 0, behavior });
-};
-
-/**
- * Returns param name from a URL.
- * @param {string} name
- */
-const getURLParams = (name) => {
-  return new URLSearchParams(window.location.search).get(name);
-};
-
-/**
- * Appends query string and returns the value in a query string format.
- * @param {string} key
- * @param {string} value
- */
-const appendURLParams = (paramName, value) => {
-  const searchParams = new URLSearchParams(window.location.search);
-  searchParams.set(paramName, value);
-  return searchParams.toString();
-};
-
-/**
- * Caching values with expiry date to the LocalStorage.
- * @param {string} key Local Storage Key
- * @param {any} value Local Storage Value
- * @param {number} ttl Time to live (Expiry Date in MS)
- */
-const setLocalItem = (key, value, ttl) => {
-  const now = new Date();
-  // `item` is an object which contains the original value
-  // as well as the time when it's supposed to expire
-  const item = {
-    value,
-    expiry: now.getTime() + ttl,
-  };
-  localStorage.setItem(key, JSON.stringify(item));
-};
-
-/**
- * Getting values with expiry date from LocalStorage that stored with `setLocalItem`.
- * @param {string} key Local Storage Key
- */
-const getLocalItem = (key) => {
-  const itemStr = localStorage.getItem(key);
-  // if the item doesn't exist, return null
-  if (!itemStr) {
-    return null;
-  }
-  const item = JSON.parse(itemStr);
-  const now = new Date();
-  // compare the expiry time of the item with the current time
-  if (now.getTime() > item.expiry) {
-    // If the item is expired, delete the item from storage
-    // and return null
-    localStorage.removeItem(key);
-    return null;
-  }
-  return item.value;
-};
-
-/**
- * Hides an HTML element when scroll down and reveals it when scroll up.
- * @param {string} id the `id` of an `HTML` element.
- * @param {number} distance An integer of pixels.
- */
-const scrollToHide = (id, distance) => {
-  const distanceDown = distance;
-  const distanceUp = distance * 2;
-  let prevScrollpos = window.pageYOffset;
-  window.onscroll = () => {
-    const currentScrollPos = window.pageYOffset;
-    if (prevScrollpos > currentScrollPos) {
-      document.getElementById(
-        id
-      ).style.transform = `translateY(${distanceDown}px)`;
-    } else {
-      document.getElementById(
-        id
-      ).style.transform = `translateY(-${distanceUp}px)`;
-    }
-    prevScrollpos = currentScrollPos;
-  };
-};
-
-// =============================== Forms =========================== //
-
-/**
- * Prevent input from typing certain keyboard chars passed as an array of chars.
- * @param {event} event input event
- * @param {Array<string>} charArr array of chars
- */
-const preventChars = (event, charArr) =>
-  charArr.includes(event.key) && event.preventDefault();
+import * as strings from "./src/strings";
+import * as numbers from "./src/numbers";
+import * as arrays from "./src/arrays";
+import * as dateAndTime from "./src/dateAndTime";
+import * as objects from "./src/objects";
+import * as html from "./src/html";
+import * as forms from "./src/forms";
+import * as window from "./src/window";
 
 module.exports = {
-  // Strings
-  capitalize,
-  truncate,
-  toggleStrNum,
-  replaceCamelCaseWithSpaces,
-  logFormattedStrings,
-  // Numbers
-  randomNumber,
-  humanFileSize,
-  formatNumber,
-  // Arrays
-  containsAll,
-  getUniqueObjs,
-  // Time
-  getTimes,
-  // Objects
-  toFormData,
-  softClone,
-  // HTML
-  getInnerHTML,
-  // Form
-  preventChars,
-  // Window
-  getScreenWidth,
-  toTop,
-  getURLParams,
-  appendURLParams,
-  setLocalItem,
-  getLocalItem,
-  scrollToHide,
+  ...strings,
+  ...numbers,
+  ...arrays,
+  ...dateAndTime,
+  ...objects,
+  ...html,
+  ...forms,
+  ...window,
 };
